- Correct eZ Publish naming, copyright dates, links to install docs etc...
git-svn-id: http://svn.projects.ez.no/ezoracle/trunk@23 436571c7-bdc5-44d1-b1ef-e55a79f94445

+++ update/database/ezpublish/4.0/dbupdate-4.0.0-to-4.0.1.sql
@@ -9,4 +9,4 @@ CREATE INDEX ezcontentobject_owner ON ezcontentobject (owner_id);
 CREATE UNIQUE INDEX ezcontentobject_remote_id ON ezcontentobject (remote_id);
 -- END: from 3.10.1
 
-ALTER TABLE ezurlalias_ml ADD alias_redirects int(11) NOT NULL default 1;
+ALTER TABLE ezurlalias_ml ADD alias_redirects number(11) NOT NULL default 1;
